Added passport jwt strategy and completed signin and destroy actions

+++ controllers/api/v1/posts_api.js
@@ -21,16 +21,23 @@ module.exports.index = async function(req,res){
 module.exports.destroy = async function(req,res){
 
     try {
-        console.log("Deleting post : ",req.params.id);
-        //If post exists then only we can delete
         let post = await Post.findById(req.params.id);
+        if(req.user.id == post.user ){
+            console.log("Deleting post : ",req.params.id);
+            //If post exists then only we can delete
+            let post = await Post.findById(req.params.id);
 
-        post.remove();
-                
-        await Comment.deleteMany({'post' : req.params.id});        
-        return res.status(200).json({
-            message : 'Post and comments deleted',
-        });
+            post.remove();
+                    
+            await Comment.deleteMany({'post' : req.params.id});        
+            return res.status(200).json({
+                message : 'Post and comments deleted',
+            });
+        }else{
+            return res.json('401',{
+                message : 'Unauthorized access',
+            })
+        }
 
     } catch (error) {
        console.log('********',err)
